requirements.txt
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.3.0
openpyxl==3.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0